requirements.txt
streamlit>=1.30.0
strands-agents>=0.1.1
mcp>=0.1.0
